Converted data types to smallest possible

+++ terrsko/Core/Src/materials.c
@@ -6,15 +6,17 @@
  */
 #include <string.h>
 #include <stdlib.h>
+#include <stdint.h>
+
 
 #include "ugui.h"
 #include "enums.h"
 #include "materials.h"
 
-int C_GRASS[4] = {C_GREEN, C_GREEN, C_DARK_GREEN, C_DARK_GREEN};
-int C_DIRT[4] = {C_BROWN, C_BROWN, C_SADDLE_BROWN, C_BROWN};
-int C_SKY[4] = {C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE};
-int C_BG_DIRT[4] = {C_DARK_BROWN, C_DARK_BROWN, C_DARK_BROWN, C_DARK_BROWN};
+uint16_t C_GRASS[4] = {C_GREEN, C_GREEN, C_DARK_GREEN, C_DARK_GREEN};
+uint16_t C_DIRT[4] = {C_BROWN, C_BROWN, C_SADDLE_BROWN, C_BROWN};
+uint16_t C_SKY[4] = {C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE, C_DEEP_SKY_BLUE};
+uint16_t C_BG_DIRT[4] = {C_DARK_BROWN, C_DARK_BROWN, C_DARK_BROWN, C_DARK_BROWN};
 
 //#define COLOR_DIRT {C_BROWN, C_BROWN, C_SADDLE_BROWN, C_BROWN}
 //#define COLOR_GRASS {C_GREEN, C_GREEN, C_DARK_GREEN, C_GREEN}
